changes in model

+++ src/heat_map/models.py
@@ -40,7 +40,7 @@ class Age(models.Model):
 
     
 class ethnicity_id(models.Model):
-    ethnicity_id = models.FloatField(default=0)
+    ethnicity_id = models.IntegerField(default=0)
     perc_white = models.FloatField(default=0)
     perc_african_american = models.FloatField(default=0)
     perc_native_american = models.FloatField(default=0)
@@ -53,8 +53,8 @@ class ethnicity_id(models.Model):
 class covid_cases(models.Model):
     test_date = models.CharField(default=0)
     county = models.CharField(default=0)
-    new_positives = models.FloatField(default=0)
-    Cumulative_num_posi = models.FloatField(default=0)
-    total_tests_day = models.FloatField(default=0)
-    cumul_tests = models.FloatField(default=0)
+    new_positives = models.IntegerField(default=0)
+    Cumulative_num_posi = models.IntegerField(default=0)
+    total_tests_day = models.IntegerField(default=0)
+    cumul_tests = models.IntegerField(default=0)
     C_Pos_C_Test = models.FloatField(default=0)
